chore: address @aorinevo's code review so that we can land
