+++ lib/command.js
@@ -1,6 +1,8 @@
 'use strict'
 
 const inspect = require('util').inspect
+const isPromise = require('./is-promise')
+const {applyMiddleware} = require('./middleware')
 const path = require('path')
 const Parser = require('yargs-parser')
 
@@ -232,24 +234,7 @@ module.exports = function command (yargs, usage, validation, globalMiddleware) {
 
       const middlewares = commandHandler.middlewares || []
 
-      innerArgv = middlewares
-        .reduce((accumulation, middleware) => {
-          if (isPromise(accumulation)) {
-            return accumulation
-              .then(initialObj =>
-                Promise.all([initialObj, middleware(initialObj)])
-              )
-              .then(([initialObj, middlewareObj]) =>
-                Object.assign(initialObj, middlewareObj)
-              )
-          } else {
-            const result = middleware(innerArgv)
-
-            return isPromise(result)
-              ? result.then(middlewareObj => Object.assign(accumulation, middlewareObj))
-              : Object.assign(accumulation, result)
-          }
-        }, innerArgv)
+      innerArgv = applyMiddleware(innerArgv, middlewares)
 
       const handlerResult = isPromise(innerArgv)
         ? innerArgv.then(argv => commandHandler.handler(argv))
@@ -448,8 +433,4 @@ module.exports = function command (yargs, usage, validation, globalMiddleware) {
   }
 
   return self
-}
-
-function isPromise(maybePromise) {
-  return maybePromise instanceof Promise
 }
