+++ lib/is-promise.js
@@ -1,0 +1,3 @@
+module.exports = function isPromise (maybePromise) {
+  return maybePromise instanceof Promise
+}
