+++ lib/middleware.js
@@ -1,3 +1,5 @@
+const isPromise = require('./is-promise')
+
 module.exports = function (globalMiddleware, context) {
   return function (callback) {
     if (Array.isArray(callback)) {
@@ -7,4 +9,25 @@ module.exports = function (globalMiddleware, context) {
     }
     return context
   }
+}
+
+module.exports.applyMiddleware = function (argv, middlewares) {
+  return middlewares
+    .reduce((accumulation, middleware) => {
+      if (isPromise(accumulation)) {
+        return accumulation
+          .then(initialObj =>
+            Promise.all([initialObj, middleware(initialObj)])
+          )
+          .then(([initialObj, middlewareObj]) =>
+            Object.assign(initialObj, middlewareObj)
+          )
+      } else {
+        const result = middleware(argv)
+
+        return isPromise(result)
+          ? result.then(middlewareObj => Object.assign(accumulation, middlewareObj))
+          : Object.assign(accumulation, result)
+      }
+    }, argv)
 }
