+++ test/command.js
@@ -1458,36 +1458,4 @@ describe('Command', () => {
       return done()
     })
   })
-
-  // addresses https://github.com/yargs/yargs/issues/1237
-  it('fails when the promise returned by the middleware rejects', (done) => {
-    const error = new Error()
-    const handlerErr = new Error('should not have been called')
-    yargs('foo')
-      .command('foo', 'foo command', noop, (yargs) => done(handlerErr), [ (yargs) => Promise.reject(error) ])
-      .fail((msg, err) => {
-        expect(msg).to.equal(null)
-        expect(err).to.equal(error)
-        done()
-      })
-      .argv
-  })
-
-  it('calls the command handler when all middleware promises resolve', (done) => {
-    const middleware = (key, value) => () => new Promise((resolve, reject) => {
-      setTimeout(() => {
-        return resolve({ [key]: value })
-      }, 5)
-    })
-    yargs('foo hello')
-      .command('foo <pos>', 'foo command', () => {}, (yargs) => {
-        yargs.hello.should.equal('world')
-        yargs.foo.should.equal('bar')
-        done()
-      }, [ middleware('hello', 'world'), middleware('foo', 'bar') ])
-      .fail((msg, err) => {
-        return done(Error('should not have been called'))
-      })
-      .argv
-  })
 })
